handlers, fixed import of packages, fixed enterContinue, fixed jsonHelper JSON path
