+++ Library/__init__.py
@@ -1,5 +1,8 @@
-from .utils import *
-from .jsonHelper import *
-from .entities import *
-from .handlers import *
-from .game import *
+# from .utils import *
+# from .jsonHelper import *
+# from .entities import *
+# from .handlers import *
+# from .game import *
+
+from . import utils, jsonHelper, entities, game
+from .handlers import messageHandlers, enemyAI, playerInput, playerAttack, magicHandlers
