+++ Library/game.py
@@ -1,4 +1,5 @@
-from Library import utils, entities, handlers
+from . import utils, entities
+from .handlers import playerInput, messageHandlers
 
 # Labels
 # TODO use these vars to provide different versions of the messages later, using data saved in something like lines.json
@@ -15,7 +16,7 @@ handleResultMessage = "Do you wish to continue on your journey? (y/n): "
 def play():
     player = entities.player()
     player.hp = 10 # TODO difficulty options, affect damage as well, affect also the enemy, both hp and damage
-    utils.intro()
+    messageHandlers.intro()
     player.name = input(playerNameMessage)
     utils.enterContinue(dangerousJourneyMessage, False, True)
     loop(player)
@@ -30,7 +31,7 @@ def loop(player):
         utils.enterContinue(tooWeak%creature, False, True)
         loop(player)
     else:
-        handlers.handlePlayerInput(player, creature)
+        playerInput.handlePlayerInput(player, creature)
     handleResult(player, creature)
     
 def handleResult(player, creature):
@@ -41,6 +42,6 @@ def handleResult(player, creature):
     if action == "y":
         loop(player)
     elif action == "n":
-        handlers.handleVictory(player)
+        messageHandlers.handleVictory(player)
     else:
         handleResult(player, creature)
