+++ Library/handlers/__init__.py
@@ -1,0 +1,8 @@
+# from .messageHandlers import *
+# from .playerInput import *
+# from .enemyAI import *
+# from .playerAttack import *
+# from .magicHandlers import *
+
+from .. import utils, jsonHelper, entities, game
+from . import messageHandlers, enemyAI, playerInput, playerAttack, magicHandlers
