+++ main.py
@@ -1,4 +1,7 @@
-import Library
+from Library import game
 
 if __name__ == "__main__":
-    Library.game.play()
+    game.play()
+
+# FIXME fix the imports tomorrow, it doesnt work !!!
+# FIXME root __init__.py
